+++ gql-core/src/test/groovy/gql/dsl/AsyncExecutionSpec.groovy
@@ -3,19 +3,18 @@ package gql.dsl
 import gql.DSL
 import gql.test.util.Queries
 import graphql.ExecutionResult
-import graphql.GraphQLException
 import graphql.schema.DataFetchingEnvironment
 import graphql.schema.GraphQLInputObjectType
 import graphql.schema.GraphQLOutputType
 import graphql.schema.GraphQLObjectType
 import graphql.schema.GraphQLSchema
 import spock.lang.Specification
-import spock.lang.Unroll
 import java.util.concurrent.CompletableFuture
 
 /**
  * @since 0.1.9
  */
+@SuppressWarnings('VariableName')
 class AsyncExecutionSpec extends Specification {
 
   void 'create an simple input type'() {
@@ -111,7 +110,6 @@ class AsyncExecutionSpec extends Specification {
     result.data.first.title == 'DR. NO'
     result.data.first.year == '1962'
 
-
     when:
     // tag::staticQueryUnchecked[]
     CompletableFuture<ExecutionResult> future2 = DSL.executeAsync(schema) {
@@ -138,5 +136,56 @@ class AsyncExecutionSpec extends Specification {
     String title
     String bond
     Integer year
+  }
+
+  void 'using execution builder'() {
+    given: 'the definition of an input and an output type'
+    GraphQLInputObjectType MailFilterType = DSL.input('MailFilter') {
+      field 'from', GraphQLString
+      field 'to', GraphQLString
+    }
+
+    GraphQLOutputType MailResult = DSL.type('Mail') {
+      field 'subject', GraphQLString
+    }
+
+    and: 'defining the schema to query'
+    GraphQLSchema schema = DSL.schema {
+      queries {
+        field('searchByFilter') {
+          type list(MailResult)
+
+          argument 'filter', MailFilterType // --> input type
+
+          fetcher { DataFetchingEnvironment env ->
+            assert env.arguments.filter.from == 'me@somedomain.com'
+            assert env.arguments.filter.to == 'you@somedomain.com'
+
+            return [[subject: "just this email here ${env.context.user}!"]]
+          }
+        }
+      }
+    }
+    and: 'the query'
+    def query = '''
+    query QueryMail($filter: MailFilter) {
+        result: searchByFilter(filter: $filter) {
+           subject
+        }
+    }
+    '''
+
+    when: 'executing the query with the required parameters'
+    CompletableFuture<ExecutionResult> future =
+      DSL.executeAsync(schema, query) {
+        withContext(user: 'John')
+        withVariables(filter: [
+          from: 'me@somedomain.com',
+          to: 'you@somedomain.com'
+        ])
+      }
+
+    then: 'we should get what we want'
+    future.get().data.result.first().subject == 'just this email here John!'
   }
 }